Move existing Core types into domain files
Move unchanged definitions from `core/types.ts` into `review-identity`, `review-comments`, `review-history`, `walkthrough`, and `generation`. Keep `core/types.ts` as the compatibility re-export, publish `./types`, and make no runtime behavior change.
