+++ core/types/review-comments.ts
@@ -1,0 +1,97 @@
+import type { MarkdownAnnotationAnchor } from '@nkzw/mdx-editor';
+import type { ReviewAuthor } from './review-history.ts';
+import type { ReviewSource } from './review-identity.ts';
+
+export type PlanCommentAuthor = {
+  avatarUrl?: string;
+  email?: string;
+  id: string;
+  name: string;
+  username?: string;
+};
+
+export type PlanCommentMessage = {
+  author: PlanCommentAuthor;
+  body: string;
+  canDelete?: boolean;
+  canEdit?: boolean;
+  createdAt: string;
+  id: string;
+  updatedAt: string;
+};
+
+export type PlanCommentThread = {
+  anchor: MarkdownAnnotationAnchor;
+  canReply?: boolean;
+  canResolve?: boolean;
+  createdAt: string;
+  createdBy: PlanCommentAuthor;
+  id: string;
+  messages: ReadonlyArray<PlanCommentMessage>;
+  resolution?: { reason: 'agent-handled' | 'anchor-removed'; resolvedAt: string };
+  status: 'open' | 'resolved';
+  updatedAt: string;
+};
+
+export type PlanReview = {
+  document: { id: string; path: string; version: string };
+  threads: ReadonlyArray<PlanCommentThread>;
+  version: 1;
+};
+
+export type PlanHandoffStatus = 'closed' | 'done';
+
+export type PullRequestReviewComment = {
+  anchor?: 'file' | 'line';
+  body: string;
+  filePath: string;
+  lineNumber?: number;
+  sectionId?: string;
+  side?: 'additions' | 'deletions';
+  startLineNumber?: number;
+  startSide?: 'additions' | 'deletions';
+  threadId?: string;
+};
+
+export type PullRequestExistingReviewComment = PullRequestReviewComment & {
+  author: ReviewAuthor;
+  canDelete?: boolean;
+  canEdit?: boolean;
+  canReplyThread?: boolean;
+  canResolveThread?: boolean;
+  id: string;
+  isOutdated?: boolean;
+  isThreadResolved?: boolean;
+  submittedAt?: string;
+  url?: string;
+};
+
+export type PullRequestGeneralComment = {
+  author: ReviewAuthor;
+  body: string;
+  canDelete?: boolean;
+  canEdit?: boolean;
+  id: string;
+  submittedAt?: string;
+  url?: string;
+};
+
+export type PullRequestGeneralCommentThread = {
+  canReply?: boolean;
+  canResolve?: boolean;
+  comments: ReadonlyArray<PullRequestGeneralComment>;
+  id: string;
+  isResolved?: boolean;
+};
+
+export type PullRequestReviewEvent = 'APPROVE' | 'COMMENT' | 'REQUEST_CHANGES';
+export type SubmitPullRequestCommentRequest = {
+  comment: PullRequestReviewComment;
+  source: Extract<ReviewSource, { type: 'pull-request' }>;
+};
+export type SubmitPullRequestReviewRequest = {
+  body?: string;
+  comments: ReadonlyArray<PullRequestReviewComment>;
+  event: PullRequestReviewEvent;
+  source: Extract<ReviewSource, { type: 'pull-request' }>;
+};
